closes #13 closes #10
The error handling is much better, though could still use some work. Need to display an appropriate message when the catalog is not obtainable.

+++ TemporaPray/Model/FileCache.swift
@@ -12,27 +12,33 @@ import os
 //TODO handle loading from disk
 public class FileCache : Codable {
     
-    var fileURLMap: [String:URL]
-    let libraryPath = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!
-    let manifestFileURL = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!.appendingPathComponent("downloadedFilesManifest.json")
+    var fileURLMap: [String:String]
+//    let libraryPath = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!
+    static let cachePath = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!
+    static let manifestFileURL = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!.appendingPathComponent("downloadedFilesManifest.json")
     
     private init() {
         fileURLMap = Dictionary()
     }
     
-   
+    
     
     private static let sharedFileCache :FileCache = {
         let fileCache = FileCache()
-        try? FileManager.default.createDirectory(at: fileCache.libraryPath, withIntermediateDirectories: true, attributes: nil)
-        let manifestFile = fileCache.libraryPath.appendingPathComponent("downloadedFilesManifest.json")
+        try? FileManager.default.createDirectory(at: cachePath, withIntermediateDirectories: true, attributes: nil)
+        
+        let manifestFile = cachePath.appendingPathComponent("downloadedFilesManifest.json")
     
         do {
             let manifestData = try? Data(contentsOf: manifestFile)
             if manifestData == nil {
                 
             } else {
-        //        fileCache.fileURLMap = try JSONDecoder().decode(Dictionary<String,URL>.self, from: manifestData!)
+                fileCache.fileURLMap = try JSONDecoder().decode(Dictionary<String,String>.self, from: manifestData!)
+                //mp3s could have been deleted in the meantime, remove any from map that don't exist.
+                
+                let newMap = fileCache.fileURLMap.filter({ !FileManager.default.fileExists(atPath: cachePath.appendingPathComponent($0.value).absoluteString)})
+                fileCache.fileURLMap = newMap
             }
             
         } catch {
@@ -45,32 +51,45 @@ public class FileCache : Codable {
     private func updateManifest() {
         do {
             let jsonOut = try JSONEncoder().encode(fileURLMap)
-            try? FileManager.default.removeItem(at: manifestFileURL)
-            try? jsonOut.write(to: manifestFileURL)
+            try? FileManager.default.removeItem(at: FileCache.manifestFileURL)
+            try? jsonOut.write(to: FileCache.manifestFileURL)
         } catch {
             os_log("Problem writing file cache manifest to disk")
         }
         
     }
+   
     
     func downloadFileFromURLAsync(urlString: String, callback: @escaping (_: Bool) -> ()) {
         //download file, save to disk, add to registry
         let url = URL(string: urlString)!
         
-//        if fileURLMap[urlString] != nil {
-//            return
-//        }
+        //check local registry first
+        if let _ = fileURLMap[urlString] {
+            callback(true)
+            return
+        
+        }
         
         let task = URLSession.shared.downloadTask(with: url) { localURL, urlResponse, error in
+            if let httpResponse = urlResponse as? HTTPURLResponse {
+                if httpResponse.statusCode == 404 {
+                    os_log("404 on attempting download of file: %@", log: .default, type: .error, urlString)
+                    callback(false)
+                    return
+                }
+            } else {
+                os_log("error connecting to server")
+                callback(false)
+                return
+            }
             if let localUrl = localURL {
-                let destinationURL = self.libraryPath.appendingPathComponent(url.lastPathComponent)
+                let destinationURL = FileCache.cachePath.appendingPathComponent(url.lastPathComponent)
                 do {
                     let fileManager = FileManager.default
                     try? fileManager.removeItem(at: destinationURL)
-//                    print(localUrlh)
-                    
                     try FileManager.default.copyItem(at: localUrl, to: destinationURL)
-                    self.fileURLMap[urlString] = destinationURL
+                    self.fileURLMap[urlString] = url.lastPathComponent
                     self.updateManifest()
                     os_log("Downloaded file: %@", log: .default, type: .info, urlString)
                     callback( true)
@@ -80,17 +99,15 @@ public class FileCache : Codable {
                     callback(false)
                 }
             }
+//            print(urlResponse)
             callback(false)
         }
         task.resume()
     }
     
-//    func downloadFileFromURLSync(url: String) {
-//
-//    }
     
-    subscript(key: String) -> URL? {
-        return  fileURLMap[key]
+    subscript(key: String) -> String? {
+        return  FileCache.cachePath.appendingPathComponent(fileURLMap[key]!).absoluteString
     }
     
     static func shared() -> FileCache {
